Create ContextAwareException
